move get_default_config to test
It is not used anywhere else.
